Refactor Diet Study Playback (#1083)
* Refactor

* Remove export from index.ts

* Restore dietScore service registraion

* Remove unused dietScoreService

* Fix spelling and type
