+++ src/features/diet-study-playback/DietStudyPlaybackCoordinator.ts
@@ -11,7 +11,6 @@ class DietStudyPlaybackCoordinator extends Coordinator {
   appCoordinator: AppCoordinator;
   navigation: NavigationType;
   contentService: IContentService;
-  dietScoreSevice: IDietScoreRemoteClient;
   dietScore: TDietScoreResponse;
 
   screenFlow: ScreenFlow = {
