+++ src/features/diet-study-playback/components/gut-score/index.tsx
