+++ src/features/diet-study-playback/components/gut-score/score-category.tsx
