+++ src/features/diet-study-playback/components/gut-score/styles.ts
