+++ src/features/diet-study-playback/components/quality-score/score.tsx
